+++ prr-core/src/prr/plans/Plan.java
@@ -3,9 +3,9 @@ package prr.plans;
 import java.io.Serializable;
 
 public interface Plan extends Serializable {
-    long calculateMessageCost(int units);
-    long calculateVoiceCommunicationCost(int units, boolean isFriend);
-    long calculateVideoCommunicationCost(int units, boolean isFriend);
+    double calculateMessageCost(double units);
+    double calculateVoiceCommunicationCost(double units, boolean isFriend);
+    double calculateVideoCommunicationCost(double units, boolean isFriend);
     //FIXME too much code repetition in subclasses?
 
 }
